refactor(mf2): Move src/stringifier/message.ts -> src/data-model/stringify.ts
