+++ packages/mf2-messageformat/src/data-model/stringify.test.ts
